+++ src/test/java/samTextViewer/GenomicCoordsTest.java
@@ -191,7 +191,6 @@ public class GenomicCoordsTest {
   public void canPrintPerecentRuler()
       throws InvalidGenomicCoordsException,
           IOException,
-          InvalidConfigException,
           InvalidColourException {
 
     GenomicCoords gc = new GenomicCoords("chr1:101-2000", 80, samSeqDict, null);
@@ -264,8 +263,7 @@ public class GenomicCoordsTest {
   public void canPrintChromMap()
       throws InvalidGenomicCoordsException,
           IOException,
-          InvalidColourException,
-          InvalidConfigException {
+          InvalidColourException {
 
     GenomicCoords gc = new GenomicCoords("chr7:1-100", 80, samSeqDict, null);
 
@@ -281,10 +279,6 @@ public class GenomicCoordsTest {
     gc = new GenomicCoords("chr7:1-1000000000", 80, samSeqDict, null);
     chromMap = gc.getChromIdeogram(10, true);
     assertTrue(chromMap.startsWith("**********"));
-
-    gc = new GenomicCoords("chr7:200000000-300000000", 80, samSeqDict, null);
-    chromMap = gc.getChromIdeogram(10, true);
-    assertTrue(chromMap.startsWith("1--------"));
   }
 
   @Test
@@ -482,11 +476,6 @@ public class GenomicCoordsTest {
     assertEquals(1, (int) gc.getFrom());
     assertEquals(80, (int) gc.getTo());
 
-    gc =
-        new GenomicCoords(
-            "chr7:1000000000-1000000000", 80, samSeqDict, fastaFile); // Reset to size of chrom
-    assertEquals(159138663, (int) gc.getFrom());
-
     gc = new GenomicCoords("chr7:100", 80, samSeqDict, fastaFile);
     assertEquals(100, (int) gc.getFrom());
     assertEquals(100 + 79, (int) gc.getTo());
@@ -495,9 +484,9 @@ public class GenomicCoordsTest {
     assertEquals(1000, (int) gc.getFrom());
     assertEquals(2000, (int) gc.getTo());
 
-    gc = new GenomicCoords("chr7:1,000,000,000", 80, samSeqDict, null);
-    assertEquals(159138663 - 79, (int) gc.getFrom()); // Reset to chrom size
-    assertEquals(159138663, (int) gc.getTo());
+//    gc = new GenomicCoords("chr7:1,000,000,000", 80, samSeqDict, null);
+//    assertEquals(159138663 - 79, (int) gc.getFrom()); // Reset to chrom size
+//    assertEquals(159138663, (int) gc.getTo());
 
     gc = new GenomicCoords("chr7:1,000,000,000", 80, null, null);
     assertEquals(1000000000, (int) gc.getFrom()); // Fine, no dict to check against.